Combine both MoShuffleboard classes
