+++ src/main/java/frc/robot/subsystem/DriveSubsystem.java
@@ -99,36 +99,35 @@ public class DriveSubsystem extends SubsystemBase {
 
         this.kinematics = new SwerveDriveKinematics(fl, fr, rl, rr);
 
-        MoShuffleboard.getInstance().ifPresent(board -> {
-            forEachSwerveModule((module) -> {
-                var group = board.driveTab
-                        .getLayout(module.getKey(), BuiltInLayouts.kList)
-                        .withSize(2, 2)
-                        .withProperties(Map.of("Label Position", "RIGHT"));
-                group.addDouble(
-                                "Drive Position (raw)",
-                                () -> module.driveMotor.getRotorPosition().getValueAsDouble())
-                        .withWidget(BuiltInWidgets.kTextView);
-                group.addDouble(
-                                "Drive Position (m)",
-                                () -> module.distEncoder.getPosition().in(Units.Meters))
-                        .withWidget(BuiltInWidgets.kTextView);
-                group.addDouble(
-                                "Drive Velocity (m_s)",
-                                () -> module.distEncoder.getVelocity().in(Units.MetersPerSecond))
-                        .withWidget(BuiltInWidgets.kTextView);
-                group.addDouble(
-                                "Turn Relative (R)",
-                                () -> module.relativeEncoder.getPosition().in(Units.Rotations))
-                        .withWidget(BuiltInWidgets.kTextView);
-                group.addDouble(
-                                "Turn Absolute (R)",
-                                () -> module.absoluteEncoder.getPosition().in(Units.Rotations))
-                        .withWidget(BuiltInWidgets.kTextView);
-            });
-
-            board.driveTab.add(this);
+        forEachSwerveModule((module) -> {
+            var group = MoShuffleboard.getInstance()
+                    .driveTab
+                    .getLayout(module.getKey(), BuiltInLayouts.kList)
+                    .withSize(2, 2)
+                    .withProperties(Map.of("Label Position", "RIGHT"));
+            group.addDouble(
+                            "Drive Position (raw)",
+                            () -> module.driveMotor.getRotorPosition().getValueAsDouble())
+                    .withWidget(BuiltInWidgets.kTextView);
+            group.addDouble(
+                            "Drive Position (m)",
+                            () -> module.distEncoder.getPosition().in(Units.Meters))
+                    .withWidget(BuiltInWidgets.kTextView);
+            group.addDouble(
+                            "Drive Velocity (m_s)",
+                            () -> module.distEncoder.getVelocity().in(Units.MetersPerSecond))
+                    .withWidget(BuiltInWidgets.kTextView);
+            group.addDouble(
+                            "Turn Relative (R)",
+                            () -> module.relativeEncoder.getPosition().in(Units.Rotations))
+                    .withWidget(BuiltInWidgets.kTextView);
+            group.addDouble(
+                            "Turn Absolute (R)",
+                            () -> module.absoluteEncoder.getPosition().in(Units.Rotations))
+                    .withWidget(BuiltInWidgets.kTextView);
         });
+
+        MoShuffleboard.getInstance().driveTab.add(this);
     }
 
     public SwerveModulePosition[] getWheelPositions() {
